% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsglm.R
\name{control}
\alias{control}
\title{Control parameters for bsglm routines}
\usage{
control(
  epsilon = 1e-08,
  maxit = 25,
  nsamples = 1000,
  nchains = 1,
  trace = FALSE
)
}
\arguments{
\item{epsilon}{convergence tolerance}

\item{maxit}{maximum number of iterations}

\item{nsamples}{number of MCMC samples}

\item{nchains}{number of MCMC chains}

\item{trace}{log execution trace?}
}
\value{
List with control parameters.
}
\description{
Specify control parameter for bsglm routines: tolerance and maximum number
of iterations for fitting, number of chains and samples for MCMC sampling,
and an execution trace flag, similar to \code{glm.control}.
}
